requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0